operating_platform/robot/robots/aloha_v1_TODO/pyproject.toml
[project]
name = "op-robot-aloha"
version = "0.1.0"
description = ""
authors = [
    {name = "Xiang Yang",email = "ryu-yang@qq.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "dora-rs-cli (>=0.3.11,<0.4.0)",
    "dora-rerun (>=0.1.0,<0.4.0)",
    "pyzmq (>=27.0.0,<28.0.0)",
]


[tool.poetry]
packages = [
    { include = "components", from = "./" },
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"